+++ submission/tests/test_tasks.py
@@ -32,18 +32,30 @@ def test_create_zendesk_ticket(mock_create_zendesk_ticket):
     assert mock_create_zendesk_ticket.call_args == mock.call(**kwargs)
 
 
-@mock.patch('submission.helpers.send_gov_notify')
-def test_task_send_gov_notify(mock_send_gov_notify):
+@mock.patch('submission.helpers.send_gov_notify_email')
+def test_task_send_gov_notify_email(mock_send_gov_notify_email):
     kwargs = {
         'subject': 'this thing',
         'reply_to': ['reply@example.com'],
         'recipients': ['to@example.com'],
         'text_body': 'Hello',
     }
-    tasks.send_gov_notify(submission_id=1, **kwargs)
+    tasks.send_gov_notify_email(submission_id=1, **kwargs)
 
-    assert mock_send_gov_notify.call_count == 1
-    assert mock_send_gov_notify.call_args == mock.call(**kwargs)
+    assert mock_send_gov_notify_email.call_count == 1
+    assert mock_send_gov_notify_email.call_args == mock.call(**kwargs)
+
+
+@mock.patch('submission.helpers.send_gov_notify_letter')
+def test_task_send_gov_notify_letter(mock_send_gov_notify_letter):
+    kwargs = {
+        'template_id': '123456',
+        'personalisation': 'Hello',
+    }
+    tasks.send_gov_notify_letter(submission_id=1, **kwargs)
+
+    assert mock_send_gov_notify_letter.call_count == 1
+    assert mock_send_gov_notify_letter.call_args == mock.call(**kwargs)
 
 
 @mock.patch('submission.helpers.send_pardot')